+++ packages/components-js/tests/e2e/puppeteer/specs/link-tile-product.e2e.ts
@@ -1,6 +1,6 @@
 import { expectA11yToMatchSnapshot, getHTMLAttributes, selectNode, setContentWithDesignSystem } from '../helpers';
 import type { Page } from 'puppeteer';
-import { Components } from '@porsche-design-system/components';
+import type { Components } from '@porsche-design-system/components/src/components';
 
 let page: Page;
 beforeEach(async () => (page = await browser.newPage()));